util/packer: Fix removing virtualbox packages

+++ util/packer/ubuntu-14.04/scripts/upgrade.sh
@@ -6,9 +6,11 @@ export DEBIAN_FRONTEND=noninteractive
 
 apt-get update
 
-# Remove the pre-installed guest additions, as the kernel install may not be compatible with them and we'll be
-# installing our own copy anyway
-apt-get remove --purge -y virtualbox-guest-dkms virtualbox-guest-utils virtualbox-guest-x11
+if [[ "${PACKER_BUILDER_TYPE}" == "virtualbox-ovf" ]]; then
+  # Remove the pre-installed guest additions, as the kernel install may not be compatible with them and we'll be
+  # installing our own copy anyway
+  apt-get remove --purge -y virtualbox-guest-dkms virtualbox-guest-utils virtualbox-guest-x11
+fi
 
 apt-get install --install-recommends linux-generic-lts-utopic \
   -y \
